# validation/results/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "MolDesign API",
  "summary": "Plataforma de dise\u00f1o molecular asistido con pipeline cient\u00edfico reproducible.",
  "description": "MolDesign expone un pipeline cient\u00edfico basado en RDKit, AutoDock Vina y scoring expl\u00edcito. La IA interpreta resultados ya calculados y no genera m\u00e9tricas qu\u00edmicas por s\u00ed misma.",
  "version": "1.0.0-alpha.2"
 },
 "paths": {
...
  "/blockchain/certificate/{molecule_id}": {
   "get": {
    "tags": [
     "blockchain"
    ],
    "summary": "Get Certificate",
    "description": "Generates and downloads a PDF certificate for the certified molecule.",
    "operationId": "get_certificate_blockchain_certificate__molecule_id__get",
    "parameters": [
     {
      "name": "molecule_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Molecule Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/evaluation/dossier/{molecule_id}/preview": {
   "post": {
    "tags": [
     "Evaluaci\u00f3n cient\u00edfica",
     "Dossier de caso"
    ],
    "summary": "Dossier del caso en PDF (inline)",
    "description": "Genera el dossier y lo devuelve para leer en pantalla.",
    "operationId": "dossier_preview_evaluation_dossier__molecule_id__preview_post",
    "parameters": [
     {
      "name": "molecule_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Molecule Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CaseProjection"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/pdf": {}
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/evaluation/dossier/{molecule_id}/package": {
   "post": {
    "tags": [
     "Evaluaci\u00f3n cient\u00edfica",
     "Dossier de caso"
    ],
    "summary": "Paquete reproducible del caso (ZIP con manifiesto y hashes)",
    "description": "Genera el paquete verificable: PDF, manifiesto, hashes y artefactos reales.",
    "operationId": "dossier_package_evaluation_dossier__molecule_id__package_post",
    "parameters": [
     {
      "name": "molecule_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Molecule Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CaseProjection"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/zip": {}
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "CaseProjection": {
    "properties": {
     "projection_version": {
      "type": "integer",
      "maximum": 1.0,
      "minimum": 1.0,
      "title": "Projection Version",
      "default": 1
     },
     "case_id": {
      "type": "string",
      "maxLength": 128,
      "minLength": 1,
      "title": "Case Id"
     },
     "case_schema_version": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 1.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Case Schema Version"
     },
     "name": {
      "type": "string",
      "maxLength": 300,
      "minLength": 1,
      "title": "Name"
     },
     "created_at": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 64
       },
       {
        "type": "null"
       }
      ],
      "title": "Created At"
     },
     "context": {
      "$ref": "#/components/schemas/ContextoCientifico"
     },
     "inputs": {
      "$ref": "#/components/schemas/InputsDeclarados"
     },
     "preflight": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/PreflightDeclarado"
       },
       {
        "type": "null"
       }
      ]
     },
     "run": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/CorridaDeclarada"
       },
       {
        "type": "null"
       }
      ]
     },
     "decisions": {
      "items": {
       "$ref": "#/components/schemas/DecisionHumana"
      },
      "type": "array",
      "maxItems": 500,
      "title": "Decisions"
     },
     "disposition": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/DisposicionDeclarada"
       },
       {
        "type": "null"
       }
      ]
     },
     "run_inputs_relation": {
      "type": "string",
      "enum": [
       "corresponde",
       "corrida_anterior",
       "desconocida"
      ],
      "title": "Run Inputs Relation",
      "default": "desconocida"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "case_id",
     "name"
    ],
    "title": "CaseProjection",
    "description": "Proyecci\u00f3n del caso: TODO lo que el backend acepta del cliente.\n\nN\u00f3tese lo que NO est\u00e1: `storage`, rutas, poses, afinidades, scores. Lo\ncient\u00edfico lo recupera el backend de sus propias fuentes, y por eso el\ndossier puede afirmar que describe lo que se ejecut\u00f3."
   },
...
   "ContextoCientifico": {
    "properties": {
     "study_kind": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 64
       },
       {
        "type": "null"
       }
      ],
      "title": "Study Kind"
     },
     "question": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 4000
       },
       {
        "type": "null"
       }
      ],
      "title": "Question"
     },
     "decision": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 4000
       },
       {
        "type": "null"
       }
      ],
      "title": "Decision"
     },
     "system_rationale": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 4000
       },
       {
        "type": "null"
       }
      ],
      "title": "System Rationale"
     },
     "controls": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 4000
       },
       {
        "type": "null"
       }
      ],
      "title": "Controls"
     },
     "assumptions": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 4000
       },
       {
        "type": "null"
       }
      ],
      "title": "Assumptions"
     },
     "uncertainties": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 4000
       },
       {
        "type": "null"
       }
      ],
      "title": "Uncertainties"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 4000
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "ContextoCientifico",
    "description": "Las preguntas del caso. TODAS opcionales a prop\u00f3sito.\n\nUn hueco es informaci\u00f3n: el dossier lo imprime como `NO DEFINIDO` en vez de\nomitir el apartado. Omitirlo dejar\u00eda al lector sin saber si nadie pregunt\u00f3 o\nsi nadie contest\u00f3."
   },
   "CorridaDeclarada": {
    "properties": {
     "task_id": {
      "type": "string",
      "maxLength": 128,
      "minLength": 1,
      "title": "Task Id"
     },
     "input_fingerprint": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Input Fingerprint"
     },
     "execution_state": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 32
       },
       {
        "type": "null"
       }
      ],
      "title": "Execution State"
     },
     "started_at": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 64
       },
       {
        "type": "null"
       }
      ],
      "title": "Started At"
     },
     "last_error": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 2000
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Error"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "task_id"
    ],
    "title": "CorridaDeclarada",
    "description": "Identidad de la corrida seg\u00fan el caso. El backend la contrasta."
   },
   "DecisionHumana": {
    "properties": {
     "control_code": {
      "type": "string",
      "maxLength": 80,
      "minLength": 1,
      "title": "Control Code"
     },
     "fingerprint": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Fingerprint"
     },
     "decision": {
      "type": "string",
      "const": "reconocida",
      "title": "Decision",
      "default": "reconocida"
     },
     "at": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 64
       },
       {
        "type": "null"
       }
      ],
      "title": "At"
     },
     "note": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 2000
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "control_code",
     "fingerprint"
    ],
    "title": "DecisionHumana",
    "description": "Una persona vio una advertencia y sigui\u00f3 adelante.\n\nAtada a su fingerprint: reconocer algo sobre unos inputs no dice nada sobre\notros. El dossier imprime la huella junto a la decisi\u00f3n para que se vea a\nqu\u00e9 hip\u00f3tesis aplicaba."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "InputsDeclarados": {
    "properties": {
     "receptor": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ReceptorDeclarado"
       },
       {
        "type": "null"
       }
      ]
     },
     "ligand": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/LigandoDeclarado"
       },
       {
        "type": "null"
       }
      ]
     },
     "config": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ConfiguracionDeclarada"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "InputsDeclarados"
   },
...
   "PreflightDeclarado": {
    "properties": {
     "fingerprint": {
      "type": "string",
      "maxLength": 200,
      "minLength": 8,
      "title": "Fingerprint"
     },
     "generated_at": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 64
       },
       {
        "type": "null"
       }
      ],
      "title": "Generated At"
     },
     "schema_version": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Schema Version"
     },
     "execution_route": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 64
       },
       {
        "type": "null"
       }
      ],
      "title": "Execution Route"
     },
     "blockers": {
      "items": {
       "type": "string",
       "maxLength": 80
      },
      "type": "array",
      "maxItems": 100,
      "title": "Blockers"
     },
     "warnings": {
      "items": {
       "type": "string",
       "maxLength": 80
      },
      "type": "array",
      "maxItems": 100,
      "title": "Warnings"
     },
     "not_evaluated": {
      "items": {
       "type": "string",
       "maxLength": 80
      },
      "type": "array",
      "maxItems": 100,
      "title": "Not Evaluated"
     },
     "receptor_label": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Receptor Label"
     },
     "ligand_label": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 4000
       },
       {
        "type": "null"
       }
      ],
      "title": "Ligand Label"
     },
     "grid_label": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Grid Label"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "fingerprint"
    ],
    "title": "PreflightDeclarado",
    "description": "Resumen del preflight que autoriz\u00f3 la corrida.\n\nEs lo que el caso guard\u00f3, no un informe nuevo: el dossier no vuelve a\ninspeccionar nada. Si el resumen y la corrida no encajan, se declara \u2014 no se\nrecalcula para tapar la discrepancia."
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}